long add

+++ problem solve/problem_set(3)/problem1.cpp
@@ -3,8 +3,8 @@ using namespace std;
 
 int main(){
 
-    int x;
-    int y;
+    long  x;
+    long  y;
     cin>>x>>y;
 
     cout<<x <<"+" <<y <<"="<<x+y <<endl;
